Add compiler option "startRules"
"startRules" is an array that holds all possible start rules of a grammar.
Thereby it's possible to omit "parseFunctionTable" from generated output
if only one startRule is possible, or to reduce the size of the array if
there is more than one startRule.
Fix "indentMultilineParts" regex to not match every whitespace but
only blanks and tabs at the beginning of a line (for correct indentation)