New single-pass JS generator backend
Introduce a new code generator which generates JS directly from the
AST. It generates similar code, except that it uses an explicit register
allocation approach instead of a stack, and it can more easily take
advantage of compile-time constants and other special cases.

It allows sharing of success and failure blocks between an AST node and
its children, eliminating some comparisons and branches. It allows
suppression of generation of match results depending on the return
context (the "discard" flag), at the expense of some cache
fragmentation.

Labeled breaks are used to exit choice and sequence operators, instead
of nested if blocks. This improves the readability of the generated
code with no measurable cost to the benchmarks.

Also:

* Rename generate-bytecode to ast-to-stackvm, and rename
  generate-javascript to stackvm-to-javascript.
* Move static generated JavaScript code from string literals in
  generate-javascript.js to a new directory for such code in
  lib/runtime.
* Rewrite the "make browser" code generator in JavaScript and introduce
  a system for bundling the runtime files.
* Move the static code which was previously in Makefile to
  lib/runtime/browser-main.js.
* In trace use fixed-width position column, and update spec accordingly.