Fixed npm & bower badges to show PEGjs version
I worked these out for my own PEGjs project :)  The shield.io example for npm is set to show the version of npm itself, which makes it hard to work out how to adapt it for your own project.  It's slightly hard to check these in the github preview window, but I'm sure they're good.